requirements.txt
blinker==1.4
certifi==2022.9.24
charset-normalizer==2.0.12
click==8.1.3
Flask==2.0.3
Flask-Cors==3.0.10
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.23.2
pg8000==1.12.5
psycopg2-binary==2.9.3
python-dotenv==0.12.0
pytz==2020.4
requests==2.27.1
sentry-sdk==1.14.0
six==1.16.0
SQLAlchemy==1.3.15
urllib3==1.26.13
Werkzeug==2.2.2
